requirements.txt
google-cloud-compute
google-cloud-resource-manager
